adapt existing code to new folder, file naming and redux convention
