+++ generators/app/templates/src/app/components/Drawer/components/DrawerMenu/index.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
-import { actionCreators as drawerActions } from '../../../../../redux/drawerHandlers';
+import { actionCreators as drawerActions } from '../../../../../redux/drawer/actions';
 
 import DrawerMenu from './layout';
