+++ generators/app/templates/src/app/screens/home/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 
-import { actionCreators as authActions } from '../../../redux/authHandlers';
+import { actionCreators as authActions } from '../../../redux/auth/actions';
 
 import Home from './layout';
