+++ generators/app/templates/src/utils/reduxUtils.js
@@ -5,9 +5,9 @@ import Immutable from 'seamless-immutable';
  * E.G:
  * stringArrayToObject(['A', 'B', 'C']) // { A: 'A', B: 'B', C: 'C' }
  */
-export function stringArrayToObject(actionsArray) {
+export function stringArrayToObject(actionsArray, namespace = '') {
   if (actionsArray.some(actionName => !actionName || typeof actionName !== 'string')) {
     throw new Error('Action names must be strings and must not be empty');
   }
-  return Immutable(actionsArray).asObject(actionName => [actionName, actionName]);
+  return Immutable(actionsArray).asObject(actionName => [actionName, `${namespace}:${actionName}`]);
 }
